requirements.txt
torch==2.7.1
torchvision==0.22.1
pillow==11.3.0
pytesseract==0.3.13
numpy==2.3.1
tqdm==4.67.1
